feat/python_downgrade: upd setup
